% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data function to convert the variable 'YEAR' to a lubridate
 package format ymd}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{(dataset)}
}
\value{
data output from function 'fecha' (dataset)
}
\description{
function to convert the variable 'YEAR' to a lubridate
 package format ymd,convert the variable 'YEAR' of character type to
a manageable date with the lubridate package
}
\examples{
\dontrun{NOAA <- eq_clean_data(data)}
}
