% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a dataframe containing cleaned NOAA (dataframe)}
}
\value{
A character vector with labels
}
\description{
Creates a label for leaflet map
}
\examples{
\dontrun{ eq_create_label(data) }
}
