% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{dataset with 'LOCATION_NAME' variable (dataset)}
}
\value{
data with nice 'LOCATION_NAME' (dataset)
}
\description{
After remove the name of the country in the 'LOCATION_NAME'
variable and the result transforms it into uppercase and lowercase
(a somewhat absurd requirement).
}
\examples{
\dontrun{NOAA <- eq_location_clean(data)}

}
