% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(eq_data, annot_col)
}
\arguments{
\item{eq_data}{a dataframe containing info on earthquakes. The data frame should be the output of 'eq_clean_data' function}

\item{annot_col}{the name of the variable containing information to show on the map.}
}
\description{
Leaftlet map of earthquakes by country
}
\examples{
\dontrun{NOAA \%>\% eq_clean_data() \%>\% dplyr::filter(COUNTRY == "USA" & lubridate::year(Date) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% eq_map(annot_col = "popup_text")}
}
