% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{fecha}
\alias{fecha}
\title{fecha}
\usage{
fecha(n)
}
\arguments{
\item{n}{the number of the record in the data set (numeric)}
}
\value{
a register with standardize
}
\description{
function to standardize the variable 'YEAR' standardizes the years so that they all
have a length of four characters
}
\examples{
\dontrun{dt.pretty_years <- mapply(fecha, 1:dim(data)[1])}

}
