% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
The function to group the elements to complement the theme
}
\examples{
\dontrun{NOAA \%>\% eq_clean_data() \%>\% dplyr::filter(COUNTRY \%in\% c("MEXICO"),
YEAR > 2000, DEATHS>0) \%>\% ggplot2::ggplot(ggplot2::aes(x = Date, y = COUNTRY,
color = DEATHS, size = as.numeric(EQ_PRIMARY)  )) + geom_timeline()}

}
