% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{theme_timeline}
\alias{theme_timeline}
\title{theme_timeline}
\usage{
theme_timeline()
}
\description{
A pretty and simple theme fot the NOAA data
}
\examples{
\dontrun{`\%>\%` <- magrittr::`\%>\%`NOAA \%>\% eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("MEXICO"), YEAR > 2000, DEATHS>0) \%>\%
ggplot2::ggplot(ggplot2::aes(x = Date, y = COUNTRY, color = DEATHS,
size = as.numeric(EQ_PRIMARY)  )) + geom_timeline() }

}
